+++ app/src/main/java/univ/soongsil/undercover/fragment/Fragment1.java
@@ -1,19 +1,22 @@
-package univ.soongsil.undercover;
+package univ.soongsil.undercover.fragment;
 
 import android.os.Bundle;
+
+import androidx.fragment.app.Fragment;
+
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
-import androidx.fragment.app.Fragment;
+import univ.soongsil.undercover.R;
 
-public class Fragment2 extends Fragment {
+public class Fragment1 extends Fragment {
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.activity_freinds_page, container, false);
+        return inflater.inflate(R.layout.activity_main_page, container, false);
 
     }
