+++ app/src/main/java/univ/soongsil/undercover/fragment/Fragment2.java
@@ -1,4 +1,4 @@
-package univ.soongsil.undercover;
+package univ.soongsil.undercover.fragment;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -7,13 +7,15 @@ import android.view.ViewGroup;
 
 import androidx.fragment.app.Fragment;
 
-public class Fragment3 extends Fragment {
+import univ.soongsil.undercover.R;
+
+public class Fragment2 extends Fragment {
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.activity_add_freinds, container, false);
+        return inflater.inflate(R.layout.activity_freinds_page, container, false);
 
     }
