+++ app/src/main/java/univ/soongsil/undercover/fragment/Fragment4.java
@@ -1,20 +1,21 @@
-package univ.soongsil.undercover;
+package univ.soongsil.undercover.fragment;
 
 import android.os.Bundle;
-
-import androidx.fragment.app.Fragment;
-
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
-public class Fragment1 extends Fragment {
+import androidx.fragment.app.Fragment;
+
+import univ.soongsil.undercover.R;
+
+public class Fragment4 extends Fragment {
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.activity_main_page, container, false);
+        return inflater.inflate(R.layout.activity_setting_actvity, container, false);
 
     }
